requirements.txt
gekko==1.0.6
matplotlib==3.8.0
numpy==1.25.2
pandas==2.1.0
plotly==5.17.0
PyYAML==6.0.1
scipy==1.11.2
streamlit==1.26.0
tqdm==4.66.1
